Remove GMP dependency since not tested
Fixes #148
